requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
Deprecated==1.2.10
dj-database-url==0.5.0
Django==3.1.4
django-cors-headers==3.5.0
django-heroku==0.3.1
django-pwa==1.0.10
djangorestframework==3.12.2
gunicorn==20.0.4
idna==2.10
nexmo==2.5.2
Pillow==8.0.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.0.0
pytz==2020.4
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
whitenoise==5.2.0
wrapt==1.12.1
